fix toggle layer on search
