+++ app/javascript/pages/map/components/menu/components/sections/datasets/index.js
@@ -1,16 +1,9 @@
 import { createElement, PureComponent } from 'react';
 import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
-import remove from 'lodash/remove';
 
 import { setModalMeta } from 'components/modals/meta/meta-actions';
-import { setMapSettings } from 'components/map-v2/actions';
 import Component from './component';
-
-const actions = {
-  setModalMeta,
-  setMapSettings
-};
 
 class DatasetsMenuContainer extends PureComponent {
   static propTypes = {
@@ -38,40 +31,13 @@ class DatasetsMenuContainer extends PureComponent {
     });
   };
 
-  onToggleLayer = (data, value) => {
-    const { activeDatasets } = this.props;
-    const { dataset, layer, iso } = data;
-    let newActiveDatasets = [...activeDatasets];
-    if (!value) {
-      newActiveDatasets = remove(newActiveDatasets, l => l.dataset !== dataset);
-    } else {
-      newActiveDatasets = [
-        {
-          dataset,
-          opacity: 1,
-          visibility: true,
-          layers: [layer],
-          ...(iso &&
-            iso.length === 1 && {
-              iso: iso[0]
-            })
-        }
-      ].concat([...newActiveDatasets]);
-    }
-    this.props.setMapSettings({
-      datasets: newActiveDatasets || [],
-      ...(value && { canBound: true })
-    });
-  };
-
   render() {
     return createElement(Component, {
       ...this.props,
-      onToggleLayer: this.onToggleLayer,
       handleRemoveCountry: this.handleRemoveCountry,
       handleAddCountry: this.handleAddCountry
     });
   }
 }
 
-export default connect(null, actions)(DatasetsMenuContainer);
+export default connect(null, { setModalMeta })(DatasetsMenuContainer);
